safari-534.53-branch/LayoutTests/sputnik/Conformance/11_Expressions/11.13_Assignment_Operators/11.13.2_Compound_Assignment/S11.13.2_A2.2_T11-expected.txt

S11.13.2_A2.2_T11

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
